+++ .gitignore
@@ -3,5 +3,5 @@
 /project/*-shim.sbt
 /project/project/
 /project/target/
-/target/
+target
 .idea
