+++ core/src/main/scala/oriented/Direction.scala
