+++ core/src/main/scala/oriented/OrientClient.scala
