+++ core/src/main/scala/oriented/Type.scala
