+++ core/src/main/scala/oriented/free/dsl/EdgeDSL.scala
