+++ core/src/main/scala/oriented/free/dsl/VertexDSL.scala
