+++ core/src/main/scala/oriented/free/interpreters/EdgeInterpreter.scala
