+++ core/src/main/scala/oriented/free/interpreters/VertexInterpreter.scala
