+++ core/src/main/scala/oriented/maps/FromMappable.scala
@@ -9,50 +9,6 @@ trait FromMappable[L, M] {
   def apply(m: M): Option[L]
 }
 
-trait LowPriorityFromMappableOption1 {
-  implicit def hconsFromMappableOption1[K <: Symbol, V, T <: HList, M]
-  (implicit wit: Witness.Aux[K], H: Lazy[MappableType[M, V]], T: FromMappable[T, M])
-  : FromMappable[FieldType[K, Option[V]] :: T, M] = new FromMappable[FieldType[K, Option[V]] :: T, M] {
-    override def apply(m: M): Option[FieldType[K, Option[V]] :: T] = T(m).map(t => field[K](H.value.get(m, wit.value.name)) :: t)
-  }
-}
-
-trait LowPriorityFromMappableOption0 extends LowPriorityFromMappableOption1 {
-  implicit def hconsFromMappableOption0[K <: Symbol, V, H <: HList, T <: HList, M]
-  (implicit wit: Witness.Aux[K], gen: LabelledGeneric.Aux[V, H], bmt: BaseMappableType[M],
-   H: Lazy[FromMappable[H, M]], T: FromMappable[T, M])
-  : FromMappable[FieldType[K, Option[V]] :: T, M] = new FromMappable[FieldType[K, Option[V]] :: T, M] {
-    override def apply(m: M): Option[FieldType[K, Option[V]] :: T] = for {
-      t <- T(m)
-    } yield {
-      val o = for {
-        n <- bmt.get(m, wit.value.name)
-        h <- H.value(n)
-      } yield gen.from(h)
-      field[K](o) :: t
-    }
-  }
-}
-
-trait LowPriorityFromMappableSeq0 extends LowPriorityFromMappableOption0 {
-  implicit def hconsFromMappableSeq0[K <: Symbol, V, H <: HList, T <: HList, M, S[_]]
-  (implicit wit: Witness.Aux[K], gen: LabelledGeneric.Aux[V, H], bmt: BaseMappableType[M],
-   H: Lazy[FromMappable[H, M]], T: FromMappable[T, M],
-   cbf: CanBuildFrom[_, V, S[V]], toSeq: S[V] => Seq[V])
-  : FromMappable[FieldType[K, S[V]] :: T, M] = new FromMappable[FieldType[K, S[V]] :: T, M] {
-    override def apply(m: M): Option[FieldType[K, S[V]] :: T] = for {
-      t <- T(m)
-    } yield {
-      val b = cbf()
-      b ++= (for {
-        n <- bmt.getAll(m, wit.value.name)
-        h <- H.value(n)
-      } yield gen.from(h))
-      field[K](b.result()) :: t
-    }
-  }
-}
-
 trait LowerPrioFromMappable {
   implicit def hnilFromMappable[M]: FromMappable[HNil, M] = new FromMappable[HNil, M] {
     override def apply(m: M): Option[HNil] = Some(HNil)
