+++ core/src/main/scala/oriented/maps/MappableType.scala
