+++ core/src/main/scala/oriented/maps/ToMappable.scala
@@ -9,34 +9,6 @@ trait ToMappable[L, M] {
   def apply(l: L): M
 }
 
-trait LowPriorityToMappableOption1 {
-  implicit def hconsToMappableOption1[K <: Symbol, V, T <: HList, M]
-  (implicit wit: Witness.Aux[K], mt: MappableType[M, V], toT: Lazy[ToMappable[T, M]])
-  : ToMappable[FieldType[K, Option[V]] :: T, M] = new ToMappable[FieldType[K, Option[V]] :: T, M] {
-    override def apply(l: FieldType[K, Option[V]] :: T): M =
-      mt.put(wit.value.name, l.head, toT.value(l.tail))
-  }
-}
-trait LowPriorityToMappableOption0 extends LowPriorityToMappableOption1 {
-  implicit def hconsToMappableOption0[K <: Symbol, V, H <: HList, T <: HList, M]
-  (implicit wit: Witness.Aux[K], gen: LabelledGeneric.Aux[V, H], mbt: BaseMappableType[M],
-   toH: Lazy[ToMappable[H, M]], toT: Lazy[ToMappable[T, M]])
-  : ToMappable[FieldType[K, Option[V]] :: T, M] = new ToMappable[FieldType[K, Option[V]] :: T, M] {
-    override def apply(l: FieldType[K, Option[V]] :: T): M =
-      mbt.put(wit.value.name, l.head.map(h => toH.value(gen.to(h))), toT.value(l.tail))
-  }
-}
-
-trait LowPriorityToMappableSeq0 extends LowPriorityToMappableOption0 {
-//  implicit def hconsToMappableSeq0[K <: Symbol, V, H <: HList, T <: HList, M]
-//  (implicit wit: Witness.Aux[K], gen: LabelledGeneric.Aux[V, H], mbt: BaseMappableType[M],
-//   toH: Lazy[ToMappable[H, M]], toT: Lazy[ToMappable[T, M]])
-//  : ToMappable[FieldType[K, TraversableOnce[V]] :: T, M] = new ToMappable[FieldType[K, TraversableOnce[V]] :: T, M] {
-//    override def apply(l: FieldType[K, TraversableOnce[V]] :: T): M =
-//      mbt.put(wit.value.name, l.head.toSeq.map(h => toH.value(gen.to(h))), toT.value(l.tail))
-//  }
-}
-
 trait LowerPrioToMappable {
 
   implicit def hnilToMappable[M](implicit mbt: BaseMappableType[M])
