+++ core/src/main/scala/oriented/maps/scalaMap.scala
