+++ core/src/main/scala/oriented/syntax/Query.scala
