+++ core/src/main/scala/oriented/syntax/syntax.scala
