+++ test/src/test/scala/TypeSpec.scala
