+++ test/src/test/scala/dsl/ClientSpec.scala
