+++ test/src/test/scala/dsl/VertexSpec.scala
