+++ test/src/test/scala/syntax/QuerySpec.scala
